docs : partydetail 파일 수정
